0.1.2-201806180025 added: Support expense type account in "from account" when adding/editing detail. fixed: some bugs fixing 新增/編輯明細資料時，「從帳戶」可選擇花費類別帳戶，用來記錄沖銷帳務，例如信用卡刷退。 錯誤修正

+++ financier/src/com/fsquirrelsoft/financier/context/ScheduleReceiver.java
@@ -25,17 +25,23 @@ public class ScheduleReceiver extends BroadcastReceiver {
                 String prefBackupDir = prefs.getString(Constants.BACKUP_DIR, ctxs.getSdFolder().getAbsolutePath());
                 Logger.d("BackupFolder: " + prefBackupDir);
                 File backupFolder = new File(prefBackupDir);
-                File dbFolder = context.getDatabasePath("fsf.db").getParentFile();
-                File prefFolder = new File(context.getFilesDir().getParent(), "shared_prefs");
-                int count = 0;
-                Logger.d("DBFolder: " + dbFolder.getAbsolutePath());
-                Logger.d("PrefFolder: " + prefFolder.getAbsolutePath());
-                count += Files.copyDatabases(dbFolder, backupFolder, now.getTime());
-                count += Files.copyPrefFile(prefFolder, backupFolder, now.getTime());
-                if (count > 0) {
-                    ctxs.setLastBackup(context, now.getTime());
+                boolean result = true;
+                if (!backupFolder.exists()) {
+                    result = backupFolder.mkdirs();
                 }
-                Files.removeOldBackups(backupFolder, now);
+                if (result) {
+                    File dbFolder = context.getDatabasePath("fsf.db").getParentFile();
+                    File prefFolder = new File(context.getFilesDir().getParent(), "shared_prefs");
+                    int count = 0;
+                    Logger.d("DBFolder: " + dbFolder.getAbsolutePath());
+                    Logger.d("PrefFolder: " + prefFolder.getAbsolutePath());
+                    count += Files.copyDatabases(dbFolder, backupFolder, now.getTime());
+                    count += Files.copyPrefFile(prefFolder, backupFolder, now.getTime());
+                    if (count > 0) {
+                        ctxs.setLastBackup(context, now.getTime());
+                    }
+                    Files.removeOldBackups(backupFolder, now);
+                }
             } catch (IOException e) {
                 throw new RuntimeException(e.getMessage(), e);
             }
